move external interactions to their own folders
