+++ src/artemis/external_interaction/callbacks/__init__.py
@@ -8,15 +8,13 @@ Callbacks used for the Artemis fast grid scan are prefixed with 'FGS'.
 from artemis.external_interaction.callbacks.fgs.fgs_callback_collection import (
     FGSCallbackCollection,
 )
-from artemis.external_interaction.callbacks.fgs.fgs_ispyb_callback import (
+from artemis.external_interaction.callbacks.fgs.ispyb_callback import (
     FGSISPyBHandlerCallback,
 )
-from artemis.external_interaction.callbacks.fgs.fgs_nexus_callback import (
+from artemis.external_interaction.callbacks.fgs.nexus_callback import (
     FGSNexusFileHandlerCallback,
 )
-from artemis.external_interaction.callbacks.fgs.fgs_zocalo_callback import (
-    FGSZocaloCallback,
-)
+from artemis.external_interaction.callbacks.fgs.zocalo_callback import FGSZocaloCallback
 
 __all__ = [
     "FGSCallbackCollection",
