+++ src/artemis/external_interaction/callbacks/fgs/fgs_callback_collection.py
@@ -1,14 +1,12 @@
 from typing import NamedTuple
 
-from artemis.external_interaction.callbacks.fgs.fgs_ispyb_callback import (
+from artemis.external_interaction.callbacks.fgs.ispyb_callback import (
     FGSISPyBHandlerCallback,
 )
-from artemis.external_interaction.callbacks.fgs.fgs_nexus_callback import (
+from artemis.external_interaction.callbacks.fgs.nexus_callback import (
     FGSNexusFileHandlerCallback,
 )
-from artemis.external_interaction.callbacks.fgs.fgs_zocalo_callback import (
-    FGSZocaloCallback,
-)
+from artemis.external_interaction.callbacks.fgs.zocalo_callback import FGSZocaloCallback
 from artemis.parameters import FullParameters
