+++ src/artemis/parameters.py
@@ -5,7 +5,7 @@ from dataclasses_json import dataclass_json
 
 from artemis.devices.eiger import DetectorParams
 from artemis.devices.fast_grid_scan import GridScanParams
-from artemis.external_interaction.ispyb_dataclass import IspybParams
+from artemis.external_interaction.ispyb.ispyb_dataclass import IspybParams
 from artemis.utils import Point3D
 
 SIM_BEAMLINE = "BL03S"
